requirements.txt
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==2.0.12
click==8.1.8
Flask==2.0.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
py==1.11.0
pytest==7.1.2
python-dotenv==0.19.0
requests==2.26.0
tmdbv3api==1.7.7
tomli==2.2.1
urllib3==1.26.20
Werkzeug==2.0.3
